fix(task): limit required separator reconstruction
